Use rpm_count_t everywhere for header data count - typedef'ed as uint32_t as that's the key size limit imposed by BDB,   relevant for RPM_BIN_TYPE - easy to change to whatever later on as it's now consistent everywhere - explicit casts where needed to avoid new warnings from signedness

+++ lib/package.c
@@ -486,7 +486,7 @@ verifyinfo_exit:
     dig->nbytes = 0;
 
     sig = memcpy(xmalloc(siglen), dataStart + info->offset, siglen);
-    (void) rpmtsSetSig(ts, info->tag, info->type, sig, info->count);
+    (void) rpmtsSetSig(ts, info->tag, info->type, sig, (size_t) info->count);
 
     switch (info->tag) {
     case RPMTAG_RSAHEADER:
